[fio extras] aktualizr-lite: Detect docker-app config changes
A common stumbling block people hit with aktualizr-lite is with
changing their docker-apps (or just enabling the first one). They
assume that by changing the config and restarting the daemon, that
the new changes will take place.

This changes detects such a change and will force an install of the
current target which will cause those changes to take place.

Signed-off-by: Andy Doan <andy@foundries.io>